// src/types/user.ts
type UserType = {
  id?: number;
  username?: string;
  email?: string;
  first_name?: string;
  last_name?: string;
  address?: string;
  phone_number?: string;
  status?: number;
  last_login_at?: Date | string;
  created_at?: Date | string;
  updated_at?: Date | string;
};

export type { UserType };